chore: rename SwiftTypeSearch to SwiftypeSearch
