+++ packages/gatsby-theme-newrelic/src/styles/SwiftypeSearchStyles.js
